requirements.txt
beautifulsoup4==4.13.4
faiss_cpu==1.11.0
fastapi==0.116.0
feedparser==6.0.11
numpy==2.3.1
pyserini==1.1.0
readability_lxml==0.8.4.1
Requests==2.32.4
sentence_transformers==5.0.0
streamlit==1.46.1
tqdm==4.67.1
